quick hack to hydrate july

+++ hydrate-timelines.py
@@ -26,19 +26,18 @@ class HydrateTimelines(object):
         file_names = glob.glob(
             os.path.join(self._user_path(user_id), '*hydrated*.json.gz'))
         for file in file_names:
-            os.remove(file)
+            # TEMPORARY HACK!!!
+            if '2020-07-hydrated.json.gz' in file:
+                os.remove(file)
 
     def hydrate(self):
         for i, user_id in enumerate(self.user_ids):
-            if i > 35:
-                break
             print('processing user {} #{}/{}...'.format(
                 user_id, i, len(self.user_ids)))
-            self._remove_hydrated(user_id)
+            # self._remove_hydrated(user_id)
             for infile in self._user_files(user_id):
-                # TODO: temporary hack
-                # if '2020-07.json.gz' in infile:
-                if True:
+                # TEMPORARY HACK!!!
+                if '2020-07.json.gz' in infile:
                     outfile = new_file_name(infile)
                     try:
                         os.remove(outfile)
